feat(powl,cmca,pddl): POWL 2.0 paper conformance + multifractal consequence allocation
Paper conformance (Kourani/Park/van der Aalst, arXiv:2602.15739v3):
- wf_to_powl: execution_flow no longer drops self-loop edges (Def 4.8 has no
  i != j exclusion; the filter silently destroyed legitimate cycles)
- wf_to_powl: convert_child always projects and recurses per Algorithm 3
  lines 8-10/18-20, instead of short-circuiting on singletons
- wf_to_powl: drop the free-choice pre-refusal gate; the paper's proofs need
  only safe+sound plus a hiding partition
- wire convert_and_verify (Theorem 1 language preservation) into both PDDL
  bridges, with max_len taken from the recomposed net's transition count

Multifractal consequence allocation:
- bcinr-cmca::cascade: N-generic, depth-generic, per-level-lens cascade with
  the allocator's two-path measure (rho descendant / 1-rho flat-to-subtree-
  leaves). rho = ONE reproduces the descendant-only behaviour exactly.
- bcinr-powl::multifractal: the powl -> cmca edge, over process_nodes preorder
- typed refusals (Cyclic, DegenerateSiblingSet, ZeroMassUnderNegativeLens)
  rather than silent saturation

Dispatch and rendering:
- process_toolkit: ready_set + dispatch_waves (antichain schedule; retires
  silent tau for free, refuses ChoiceGraph/DoRedo as not statically schedulable)
- process_toolkit: process_to_mermaid_annotated renders derived waves and
  consequence mass; process_to_mermaid left untouched

Soundness fixes:
- ground/monitors: remove unsound Within/AlwaysWithin monitors in favour of
  an honest None; add pending_is_satisfied_at_finalize so TrajectoryPolicy
  stops treating an Always monitor's terminal Pending as a violation
- cognitive: surface the hierarchical Refusal instead of .ok()-ing it

Verified by differential oracle: cascade and the certified branchless
allocator agree to 6.7e-4 max leaf deviation across q in {1, 0, -1} on their
pinned shared slice. Residual is exp2/log2 vs repeated saturating_mul.

Known open: cascade refuses q=2 (DegenerateSiblingSet) where the allocator's
log-domain kernel survives m_min underflow; ground_v2::legacy_action still
drops When/Numeric effects; falsification_adversarial has 11 failures that
predate this branch.

+++ crates/bcinr-powl/tests/usecase_swarm_05_deadline_miss_detection.rs
@@ -45,11 +45,11 @@ fn execute(ast: &PowlAstNode<'_>, run_id: u64) -> (PowlTape, PowlRunState, OcelL
             let op_idx = bits.trailing_zeros();
             bits &= bits - 1;
             // Record with a duration approximation: each op takes ~10 time units
-            log.record_op_fired(run_id, op_idx, ticks as u32 * 10, 10).unwrap();
+            log.record_op_fired(run_id, op_idx, ticks * 10, 10).unwrap();
             op_trace |= 1u64 << op_idx;
         }
     }
-    log.record_run_sealed(run_id, op_trace, ticks as u32 * 10).unwrap();
+    log.record_run_sealed(run_id, op_trace, ticks * 10).unwrap();
     (tape, state, log, ticks)
 }
 
@@ -146,7 +146,11 @@ fn test_partial_order_deadline_validation_fails() {
 
     // Expect deadline violation.
     match result {
-        ConformanceResult::DeadlineViolation { actual_time, deadline, .. } => {
+        ConformanceResult::DeadlineViolation {
+            actual_time,
+            deadline,
+            ..
+        } => {
             assert!(
                 actual_time > deadline,
                 "violation: {} > {}",
@@ -154,10 +158,7 @@ fn test_partial_order_deadline_validation_fails() {
                 deadline
             );
         }
-        _ => panic!(
-            "expected deadline violation, got {:?}",
-            result
-        ),
+        _ => panic!("expected deadline violation, got {:?}", result),
     }
 }
 
@@ -227,7 +228,11 @@ fn test_nested_sequence_deadline_violation() {
     // Validate and expect deadline violation.
     let result = log.validate_against_tape(&tape);
     match result {
-        ConformanceResult::DeadlineViolation { actual_time, deadline, .. } => {
+        ConformanceResult::DeadlineViolation {
+            actual_time,
+            deadline,
+            ..
+        } => {
             assert!(
                 actual_time > deadline,
                 "actual_time {} must exceed deadline {}",
@@ -321,10 +326,7 @@ fn test_deadline_logic_no_external_api_calls() {
     );
 
     // Log must have recorded the ops.
-    assert!(
-        !log.events().is_empty(),
-        "log should record events"
-    );
+    assert!(!log.events().is_empty(), "log should record events");
 }
 
 /// Test 7: ConformanceResult provides diagnostic information
